feat: init Logic track
I'll be using Logic to record the more linear dialogue tracks (rather than the procedural stuff happening in Ableton. I'm not sure if the files are going to end up getting too big, but if so, I suppose and I can add the audio files folder to the gitignore. We'll see.